fix: remove the OpenSSL dependency for Linux
Since we no longer use `http_transport_socket` on Linux there is no
need to depend on OpenSSL directly anymore. We use whatever the present
`libcurl` implementation uses as its TLS implementation.

Also moved the dependency discovery for Android (which we currently do
not use) to `crashpad-util` to isolate it to its usage.