Fix capture_context_win_arm64 in standalone clang build
clang-cl does not understand armasm64 assembler syntax. The existing
workaround for crashpad_is_in_chromium, which has used clang-cl
exclusively for quite some time, was to use a checked-in prebuilt .obj
produced by armasm64, rather than assembling the .asm during the build.
This same .obj can be used to support the standalone clang build.

As Crashpad on Windows is now clang-primary (with an eye toward becoming
clang-only), this is not a long-term solution. The existing armasm64
source will need to be translated to an assembler dialect that clang-cl
understands.

Bug: 399587142, 384682775
Change-Id: Ieeb3c1c54c954997514a3fecb5c362b7d4d7d13d
Reviewed-on: https://chromium-review.googlesource.com/c/crashpad/crashpad/+/6309923
Reviewed-by: Nico Weber <thakis@chromium.org>
Commit-Queue: Mark Mentovai <mark@chromium.org>