ios: Support intermediate dump extra_memory_ranges.
The extra_memory_ranges stored in the intermediate dump and used by
gwp-asan via the ProcessSnapshot are not needed in the minidump. If this
memory were stored in the minidump with gwp-asan, this would double the
size of every iOS minidump.

Instead, add a second bag that is stored in the intermediate dump but
not written to the minidump. Expose an iOS only ProcessSnapshot API that
can be used by the gwp-asan analyzer to use the extra_memory_ranges without increasing the minidump file size.

Change-Id: Ibf7b7fb89cda0a829727c02e5118dc2f6423769f
Reviewed-on: https://chromium-review.googlesource.com/c/crashpad/crashpad/+/6351477
Reviewed-by: Joshua Peraza <jperaza@chromium.org>
Commit-Queue: Justin Cohen <justincohen@chromium.org>