build: Add tvOS support to setup_ios_gn.py
This is similar to https://crrev.com/c/6328438; the idea is to generate
out/<config>-appletvos and out/<config>-appletvsimulator directories
that set `target_platform` to "tvos".

`target_platform` is also being set for other iOS and Mac Catalyst, in
which case it is a no-op.

Bug: 405140658
Change-Id: I61626018504a2272210eb5895f2cf064a31410dc
Reviewed-on: https://chromium-review.googlesource.com/c/crashpad/crashpad/+/6421721
Commit-Queue: Raphael Kubo da Costa <kubo@igalia.com>
Reviewed-by: Justin Cohen <justincohen@chromium.org>