Update Fuchsia toolchain pin to use `integration`
Previously crashpad had pinned the Fuchsia toolchain using the `latest`
tag, which is not correct, since we use `latest` to tag the newest green
build of the toolchain, which has not been fully vetted yet.

Instead, testing should use the `integration` revision, which is the
current in-tree toolchain, that has been qualified to use when building
Fuchsia.

Bug: 408065472
Change-Id: I2f13e20223082fe9143ee8036d597cbe3277cd95
Reviewed-on: https://chromium-review.googlesource.com/c/crashpad/crashpad/+/6428398
Commit-Queue: Mark Mentovai <mark@chromium.org>
Reviewed-by: Roland McGrath <mcgrathr@chromium.org>
Reviewed-by: Mark Mentovai <mark@chromium.org>